+++ fixtures/trait-methods/tests/bindings/test.swift
@@ -6,6 +6,7 @@ assert(String(reflecting: m) == "TraitMethods { val: \"yo\" }")
 
 // eq
 assert(m == TraitMethods(name: "yo"))
+assert(m != TraitMethods(name: "foo"))
 
 // hash
 var set: Set = [TraitMethods(name: "yo")]