+++ uniffi_bindgen/src/interface/function.rs
@@ -249,6 +249,9 @@ pub trait Callable {
     fn return_type(&self) -> Option<Type>;
     fn throws_type(&self) -> Option<Type>;
     fn is_async(&self) -> bool;
+    fn takes_self(&self) -> bool {
+        false
+    }
     fn result_type(&self) -> ResultType {
         ResultType {
             return_type: self.return_type(),
@@ -317,6 +320,10 @@ impl<T: Callable> Callable for &T {
 
     fn is_async(&self) -> bool {
         (*self).is_async()
+    }
+
+    fn takes_self(&self) -> bool {
+        (*self).takes_self()
     }
 }