+++ uniffi_udl/src/converters/callables.rs
@@ -126,6 +126,7 @@ impl APIConverter<ConstructorMetadata> for weedle::interface::ConstructorInterfa
             name: String::from(attributes.get_name().unwrap_or("new")),
             // We don't know the name of the containing `Object` at this point, fill it in later.
             self_name: Default::default(),
+            is_async: attributes.is_async(),
             // Also fill in checksum_fn_name later, since it depends on object_name
             inputs: self.args.body.list.convert(ci)?,
             throws,